revised code reorder some fct and updated readme
