+++ usecases/MinimumWorkingExample/upload_scripts/doit_create_types.py
@@ -6,45 +6,13 @@ import logging
 from lebedigital.openbis.interbis import Interbis
 from pybis.entity_type import SampleType
 
-from upload_scripts.doit_upload import _read_metadata, _reformat_sample_dict
+from upload_scripts.doit_upload import _read_metadata
 
 conv_dict = {
     str: 'VARCHAR',
     float: 'REAL',
     int: 'INTEGER',
 }
-
-
-def _create_union_yaml(yaml_directory_path: Path, output_path: Union[Path, None], mixture_code: str,
-                       defaults_dict: dict) -> dict:
-    union_dict = defaults_dict
-
-    for file in os.scandir(yaml_directory_path):
-        with open(file, "r") as stream:
-            current_dict = {key: val for key, val in dict(yaml.safe_load(stream)).items()
-                            if key not in union_dict.keys()}
-            current_dict = {f"{mixture_code}.{key}": [conv_dict[type(val)], key, key]
-                            for key, val in current_dict.items()}
-        union_dict = dict(current_dict, **union_dict)
-
-    if output_path:
-        output_file = Path(output_path, "mixfile_union.yaml")
-        with open(output_file, "w") as output:
-            output.write(yaml.dump(union_dict))
-
-    return union_dict
-
-
-def _create_logfiles(mixture_sample_type: Union[SampleType, str], emodul_sample_type: Union[SampleType, str],
-                     logging_path: Union[Path, str]):
-    with open(Path(logging_path, 'mixture_sample_type.yaml'), 'w') as file:
-        # file.write(yaml.dump({'Sample Type': mixture_sample_type}))
-        print(mixture_sample_type, file=file)
-
-    with open(Path(logging_path, 'emodul_sample_type.yaml'), 'w') as file:
-        # file.write(yaml.dump({'Sample Type': emodul_sample_type}))
-        print(emodul_sample_type, file=file)
-
 
 def create_required_sample_types(mixture_directory_path: Union[Path, str],
                                  emodul_directory_path: Union[Path, str],
@@ -70,7 +38,7 @@ def create_required_sample_types(mixture_directory_path: Union[Path, str],
         defaults_dict=default_props)
 
     mixture_sample_type = o.create_sample_type(
-        sample_code="EXPERIMENTAL_STEP_" + config['mixture_prefix'],
+        sample_code=f"EXPERIMENTAL_STEP_{config['mixture_prefix']}",
         sample_prefix=config['mixture_prefix'],
         sample_properties=mix_union_dict,
     )
@@ -79,12 +47,12 @@ def create_required_sample_types(mixture_directory_path: Union[Path, str],
     metadata_path = Path(emodul_directory_path, os.fsdecode(os.listdir(emodul_directory_path)[0]))
 
     # CREATING EMODUL SAMPLE TYPE from single yaml (all equal)
-    emodul_metadata = _read_metadata(metadata_path, "EXPERIMENTAL_STEP_" + config["emodul_prefix"], default_props)
+    emodul_metadata = _read_metadata(metadata_path, f"EXPERIMENTAL_STEP_{config['emodul_prefix']}", default_props)
     emodul_metadata_type_dict = _reformat_sample_dict(emodul_metadata)
 
     # Creating the emodul sample type with the formatted dict
     emodul_sample_type = o.create_sample_type(
-        sample_code="EXPERIMENTAL_STEP_" + config['emodul_prefix'],
+        sample_code=f"EXPERIMENTAL_STEP_{config['emodul_prefix']}",
         sample_prefix=config['emodul_prefix'],
         sample_properties=emodul_metadata_type_dict,
     )
@@ -95,3 +63,50 @@ def create_required_sample_types(mixture_directory_path: Union[Path, str],
                      logging_path=logging_path)
 
     o.logout()
+
+def _create_union_yaml(yaml_directory_path: Path, output_path: Union[Path, None], mixture_code: str,
+                       defaults_dict: dict) -> dict:
+    union_dict = defaults_dict
+
+    for file in os.scandir(yaml_directory_path):
+        with open(file, "r") as stream:
+            current_dict = {key: val for key, val in dict(yaml.safe_load(stream)).items()
+                            if key not in union_dict.keys()}
+            current_dict = {f"{mixture_code}.{key}": [conv_dict[type(val)], key, key]
+                            for key, val in current_dict.items()}
+        union_dict = dict(current_dict, **union_dict)
+
+    if output_path:
+        output_file = Path(output_path, "mixfile_union.yaml")
+        with open(output_file, "w") as output:
+            output.write(yaml.dump(union_dict))
+
+    return union_dict
+
+def _reformat_sample_dict(loaded_dict: dict):
+    conv_dict = {
+        str: 'VARCHAR',
+        float: 'REAL',
+        int: 'INTEGER',
+    }
+
+    output_dict = defaultdict(lambda: ["NA", "NA", "NA"])
+    # output_dict['$name'] = ['VARCHAR', 'Name', 'Name']
+
+    for key, val in loaded_dict.items():
+        val = "" if val is None else val
+        output_dict[key.lower()] = [conv_dict[type(val)], key.split('.')[-1], key.split('.')[-1]]
+
+    return dict(output_dict)
+
+def _create_logfiles(mixture_sample_type: Union[SampleType, str], emodul_sample_type: Union[SampleType, str],
+                     logging_path: Union[Path, str]):
+    with open(Path(logging_path, 'mixture_sample_type.yaml'), 'w') as file:
+        # file.write(yaml.dump({'Sample Type': mixture_sample_type}))
+        print(mixture_sample_type, file=file)
+
+    with open(Path(logging_path, 'emodul_sample_type.yaml'), 'w') as file:
+        # file.write(yaml.dump({'Sample Type': emodul_sample_type}))
+        print(emodul_sample_type, file=file)
+
+
